+++ src/main/java/madsciencemod/cards/Tune.java
@@ -21,7 +21,6 @@ public class Tune extends AbstractMadScienceCard {
     public static final String[] EXTENDED_DESCRIPTION = cardStrings.EXTENDED_DESCRIPTION;
     private static final int COST = 0;
     private static final int AMOUNT = 2;
-    private static final int UPGRADE_AMOUNT = 1;
     private static final CardType TYPE = CardType.SKILL;
     private static final CardRarity RARITY = CardRarity.UNCOMMON;
     private static final CardTarget TARGET = CardTarget.SELF;
@@ -55,7 +54,9 @@ public class Tune extends AbstractMadScienceCard {
     public void upgrade() {
         if (!this.upgraded) {
             upgradeName();
-            this.upgradeMagicNumber(UPGRADE_AMOUNT);
+            this.exhaust = false;
+            this.rawDescription = UPGRADE_DESCRIPTION;
+            this.initializeDescription();
         }
     }
 }